fix for shitty build system

+++ src/raylib.c
@@ -601,7 +601,7 @@ RLCAPI int GetRandomValue(int min, int max) {
         max = min;
         min = tmp;
     }
-    return (int)((float)rand() / (float)(RAND_MAX + 1) * (float)(max - min + 1)) + min;
+    return (int)((float)((double)rand() / (double)(RAND_MAX + 1)) * (float)(max - min + 1)) + min;
 }
 
 RLCAPI void SetRandomSeed(unsigned int seed) {
